Move fuzzy_match_timezone and TimezoneMatches from bios/misc to common/time.
