+++ sandpiper/common/time.py
@@ -1,13 +1,16 @@
+from dataclasses import dataclass
 import datetime as dt
 import re
-from typing import Optional, Union, cast
+from typing import *
 
+from fuzzywuzzy import fuzz, process as fuzzy_process
 import pytz
 import tzlocal
 
 __all__ = [
     'TimezoneType', 'time_format', 'parse_time', 'parse_date', 'format_date',
-    'utc_now', 'localize_time_to_datetime'
+    'utc_now', 'localize_time_to_datetime', 'TimezoneMatches',
+    'fuzzy_match_timezone'
 ]
 
 TimezoneType = Union[pytz.tzinfo.StaticTzInfo, pytz.tzinfo.DstTzInfo]
@@ -162,3 +165,46 @@ def localize_time_to_datetime(time: dt.time,
     basis_time = dt.datetime(now_basis.year, now_basis.month, now_basis.day,
                              time.hour, time.minute)
     return basis_tz.localize(basis_time)
+
+
+@dataclass
+class TimezoneMatches:
+    matches: List[Tuple[str, int]] = None
+    best_match: Optional[TimezoneType] = False
+    has_multiple_best_matches: bool = False
+
+
+def fuzzy_match_timezone(
+        tz_str: str, best_match_threshold=75, lower_score_cutoff=50, limit=5
+) -> TimezoneMatches:
+    """
+    Fuzzily match a timezone based on given timezone name.
+
+    :param tz_str: timezone name to fuzzily match in pytz's list of timezones
+    :param best_match_threshold: Score from 0-100 that the highest scoring
+        match must be greater than to qualify as the best match
+    :param lower_score_cutoff: Lower score limit from 0-100 to qualify matches
+        for storage in ``TimezoneMatches.matches``
+    :param limit: Maximum number of matches to store in
+        ``TimezoneMatches.matches``
+    """
+
+    # I think partial_token_sort_ratio provides the best experience.
+    # The regular token_sort_ratio just feels weird because it doesn't support
+    # substrings. Searching "Amst" would pick "GMT" rather than "Amsterdam".
+    # The _set_ratio methods are totally unusable.
+    matches: List[Tuple[str, int]] = fuzzy_process.extractBests(
+        tz_str, pytz.common_timezones, scorer=fuzz.partial_token_sort_ratio,
+        score_cutoff=lower_score_cutoff, limit=limit
+    )
+    tz_matches = TimezoneMatches(matches)
+
+    if matches and matches[0][1] >= best_match_threshold:
+        # Best match
+        tz_matches.best_match = pytz.timezone(matches[0][0])
+        if len(matches) > 1 and matches[1][1] == matches[0][1]:
+            # There are multiple best matches
+            # I think given our inputs and scoring algorithm, this shouldn't
+            # ever happen, but I'm leaving it just in case
+            tz_matches.has_multiple_best_matches = True
+    return tz_matches
